+++ app/src/main/java/com/example/jelenazivanovic/weatherforecastappmwp/retrofitmountaintview/models/WeatherInfo.java
@@ -16,27 +16,27 @@ public class WeatherInfo {
     @SerializedName("weather")
     private ArrayList<WeatherDetail> weatherDetail = new ArrayList<>();
     @SerializedName("pressure")
-    private double pressure;
+    private float pressure;
     @SerializedName("humidity")
-    private int humidity;
+    private float humidity;
     @SerializedName("speed")
-    private double windSpeed;
+    private float windSpeed;
     @SerializedName("deg")
-    private double windDirection;
+    private float windDirection;
 
-    public double getWindSpeed() {
+    public float getWindSpeed() {
         return windSpeed;
     }
 
-    public void setWindSpeed(double windSpeed) {
+    public void setWindSpeed(float windSpeed) {
         this.windSpeed = windSpeed;
     }
 
-    public double getWindDirection() {
+    public float getWindDirection() {
         return windDirection;
     }
 
-    public void setWindDirection(double windDirection) {
+    public void setWindDirection(float windDirection) {
         this.windDirection = windDirection;
     }
 
@@ -55,19 +55,19 @@ public class WeatherInfo {
         this.weatherDetail = weatherDetail;
     }
 
-    public double getPressure() {
+    public float getPressure() {
         return pressure;
     }
 
-    public void setPressure(double pressure) {
+    public void setPressure(float pressure) {
         this.pressure = pressure;
     }
 
-    public int getHumidity() {
+    public float getHumidity() {
         return humidity;
     }
 
-    public void setHumidity(int humidity) {
+    public void setHumidity(float humidity) {
         this.humidity = humidity;
     }
 }
